+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add feature to add change summaries when saving or publishing
 - Improve library and script assignment UI by separating out global vs local
 - Improve navigatio and side menu management UI
+- Add back navigation and side nav settings
 
 ## 1.6.0